Started styling for the dragovich home page
Styled the images, the nav, the footer. Still not entirely complete. I have made a few changes to the html structure for the home page so be sure to check those and make sure the structure for the child pages all match that. I will add the css to a stylesheet once it is complete.